requirements.txt
llama-cpp-python==0.2.33
llama-index==0.9.39
transformers==4.37.1
deepeval==0.20.55
pytest==8.0.0
git+https://github.com/openai/whisper.git
